[feature] whitelist host instead of basic auth
Something, maybe PivotalTracker or Rails itself, is stripping the
prepended auth credentials from the webhook URL, so we can't access them
to validate the request. This seems like it must be solveable, but I
haven't had any luck in ~1hr of timeboxed efforts. Whitelisting the
host should minimize malicious requests sufficiently for now, and we can
revisit the auth issue if and when that becomes a priority.